tidy: Minor cleanup of licenseck.py (#40585)
- Remove unused accepted Copyright line `See
http://rust-lang.org/COPYRIGHT`. This URL now yields a 404, and we don't
seem to use this line anywhere in the project so we can remove it. Note:
We should probably use a regex anyway, so we can match the whole
required expression, but thats a different topic.
- Remove unused `licenses_dep_toml` table. We use cargo deny for
checking the licenses of our rust dependencies.

Testing: Changing tidy is covered by `./mach test-tidy` in CI

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>