libservo: Add `WebView::can_go_forward` and `WebView::can_go_back`
These two new API method make it easier to enable and disable the back
and forward buttons in the browser UI. In addition, they are used to
enable and disable the context menu items for "Back" and "Forward". This
is something that is much, much easier to do in the embedding layer,
rather than in script.

In addition, the new methods are wired up to the buttons in Servoshell.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>