script: Correctly parse spki and pkcs8 of ECDH
We currently parse the `parameter` field of the `algorithm`
AlgorithmIdentifier field of `spki` and `pkcs8` of ECDH keys as an
object identifier. However, according to WebCrypto API specification, we
should parse it as an `ECParameters` ASN.1 type defined in [RFC5480].

```plaintext
ECParameters ::= CHOICE {
  namedCurve         OBJECT IDENTIFIER
  -- implicitCurve   NULL
  -- specifiedCurve  SpecifiedECDomain
}
```

Although our current implementation works fine since `ECParameters` is
defined as a `CHOICE` type with an object identifier as the only option,
it is technically incorrect. This patch fixes this mistake.

Testing: No behavioral change. Existing tests suffice.

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>