CI: Limit github runner mac to macos-arm-unit-tests only on merge queue (#40617)
Based on the chat here:
https://github.com/servo/servo/issues/40596#issuecomment-3528514438 we
limit mac-arm (those are bottlenecks as they are not self hosted) to
merge queue (so we know that main is good so we can skip run there) and
we do not run MSRV check there to further limit uses.

Testing: None, because it's just infra.
Temp fixes: #40596

Signed-off-by: sagudev <16504129+sagudev@users.noreply.github.com>