script: Issue resource timing for stylesheets before the "load" event (#40618)
Tests seem to assume that resource timing is available during the "load"
event. Doing so causes some tests to pass and others to fail, which
appear to have been false passes.

This is a preparatory change for asynchronous stylesheet parsing, when
resource timing entires must be added before sending stylesheets to
another thread for parsing.

Testing: This change updates WPT test results.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>