html: Handle new `DurationChanged` event by <media>
A new event `PlayerEvent::DurationChanged` has been added to
the media engine to simplify handling changes in the duration
of a media stream (sometimes the duration could be inaccurate
or even determined by the media engine with significant delay
after the initial metadata was ready).

This new event eliminates the need to check conditions for
two consecutive `MetadataUpdated` events (the initial metadata
and the metadata with the changed duration) during the loading
phase.

servo-media:
- Add new `PlayerEvent::DurationChanged` event (https://github.com/servo/media/pull/461)

Testing: No expected changes in test results

Fixes: #https://github.com/servo/servo/issues/40626

Signed-off-by: Andrei Volykhin <andrei.volykhin@gmail.com>