Support `calc()` values in CSS Grid (#34846)
Depends on:

- https://github.com/servo/stylo/pull/104
- https://github.com/servo/servo/pull/34926
- https://github.com/servo/servo/pull/34927
- https://github.com/servo/servo/pull/34948

In addition to that a `resolve_calc_value` function has been added which
resolves calc values during the layout process once a percentage
resolution basis is available.

~~There is 1 newly failing test and 1 newly failing subtest here.~~
These issues have now been fixed.

There are 8 new subtest failures in
`css/css-grid/grid-definition/grid-minimum-contribution-with-percentages.html`.
These are genuine failures, but are unrelated to the calc
implementation. The calc implementation is just exposing a pre-existing
bug around percentage resolution that also now correctly also applies to
`calc()` values containing percentages. The fix for would best be done
in a followup as it requires teaching Taffy about "compressible replaced
elements". (update: I have this
[implemented](https://github.com/servo/servo/pull/34948) but in the
interest of keeping it to one feature per PR I have not included it
here)

---
<!-- Thank you for contributing to Servo! Please replace each `[ ]` by
`[X]` when the step is complete, and replace `___` with appropriate
data: -->
- [x] `./mach build -d` does not report any errors
- [x] `./mach test-tidy` does not report any errors. It doesn't like the
git dependency on Taffy.
- [x] There are tests for these changes OR

---------

Signed-off-by: Nico Burns <nico@nicoburns.com>