Implement `self.reportError`
This web API is alternative API to `throw e`, which
is why we can reuse a lot of the existing machinery.

The one testcase that isn't passing yet is because
it reports an empty `TypeError`. The current logic
in `ErrorInfo` only retrieves the message data, but
doesn't include the type of the exception. For that,
we need to use `(*report)._base.errorNumber` and map
that back to the original type codes. However, deferring
that to a follow-up as that requires some more work
in mozjs.

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>