servodriver: Shut down Servo elegantly when all tests finish (#40455)
Previously, we always kill the instance when tests finish instead of
normally shutting down. That blocked clean-ups works, such as Code
Coverage we wanted to add, and other things I'm not aware of.

Testing: We have tested for a week and checked the logs of CI to confirm
it works normally now.

---------

Signed-off-by: Euclid Ye <yezhizhenjiakang@gmail.com>