libservo: Trigger a rendering update when waiting for paint readiness in WebView unit tests
Some unit tests attempt to wait for the WebView to be paint ready before
executing. The previous code would wait for a frame ready message from
the WebView before running, but this was subject to timing issues due to
the frame being sent before the load complete message. This changes
makes it so that the WebView explicitly triggers another rendering
update and then waits for a frame. It should be guaranteed that the next
frame will be one that reflects load completeness.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>