mach: Normalize the `--profile` argument to build commands
The current logic treats the explict version of release and dev build
specification i.e the `--profile release` and `--profile dev` options as
custom profiles. This breaks other logic that explictly check for
release and dev profile types, like the android build.

Normalize the `--profile` argument so that the strings `release` and
`dev` are treated correctly as `BuildType.Kind.RELEASE` and
`BuildType.Kind.DEV` respectively.

Fixes: #40752

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>