paint: Do trigger animations when a non-animation paint happens (#40433)
Otherwise the `RefreshObsever` can get wedged. This does mean that we
will overpaint a bit when processing input events and animating, but we
could integrate scroll / zoom event handling with the animation loop to
eliminate this overpaint.

Testing: This kind of fix for touch interaction is currently very hard
to test as it
relies heavily on the order that operations happen between Servo and the
`RefreshDriver`. It is easy to reproduce in interactive usage though and
this
change fixes it.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>