servoshell: Move `Minibrowser` to headed window
Instead of storing the `Minibrowser` egui struct in the
`RunningAppState` which is used for both headed and headless windows,
move it to the headed window implementation. This addresses a lot of
bits of the code that would need to test the existance of the
`Minibrowser` even when it should have always existed.

In addition, event handling for the headed window is moved into the
headed window implementation itself, which eliminates the need to expose
so many methods in `WindowPortsMethods`.

One downside is that it is impossible to support tracing of `AppEvent`s
as accessibilty events are not cloneable. This seems like a small price
to pay for this cleanup though.

This is the first step toward supporting multiple windows in servoshell.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>