layout: Preserve cached intrinsic sizes of undamaged descendants (#40795)
We were previously clearing the cached intrinsic inline sizes when the
element, an ancestor, or a child had relayout damage. However, intrinsic
sizes don't depend on ancestors for the most part, so now we will only
clear the cache if the relayout damage is in the element or a
descendant.

Testing: Not needed, no change in behavior

Signed-off-by: Oriol Brufau <obrufau@igalia.com>
Co-authored-by: Martin Robinson <mrobinson@igalia.com>