script: Improve spec-compliance of Performance resource timing entries (#40811)
The current implementation doesn't match the specification and is broken
in a way that makes nytimes.com endlessly attempt to enqueue resource
timing entries, making the page completely unresposive to user input.
This PR makes some surgical changes that fix the NYTimes behaviour and
add some comments to make it easier to further improve the
implementation in the future.

Testing: Various WPT that used to timeout now fail instead.
Fixes: #40808

---------

Signed-off-by: Josh Matthews <josh@joshmatthews.net>