servoshell: Remove code duplication in `EventsLoop` and rename to `ServoEventLoop`
It used to be that the winit event loop for different platforms was
initialized differently. That isn't the case any longer, so we do not
need platform-specific code here.

In addition, I think that the slightly-odd `EventsLoop` name was chosen
to not conflict with winit's `EventLoop`. Instead I've renamed this to
`ServoEventLoop` and removed one level of nesting in the data structure,
which simplifies things a bit.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>