servoshell: Rename `Minibrowser` to `Gui` (#40816)
MiniBrowser was considered at some point for the name of servoshell, but
was discarded. The name of the egui portions of servoshell are a
holdover from this. This change renames this struct to `Gui` to better
reflect what it does.

Testing: This just renames some data structures, so should not need new
tests.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>