ohos CI: Try improving scenario test stability
Previously enabling the performance mode helped
preventing time-outs in the speedometer test on hos devices.
This functionality was extracted into a seperate python library,
and we try and apply it to the scenario tests too.
Internally this use `hdc shell power-shell` to set the
- performance mode
- screen timeout
- explicit wakeup of the device

After the test / on exception, the changes performance mode is disabled again (hence the `with``.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>