Allow webdriver to automatically accept custom protocol handlers
This reads the `protocol_handler_automation_mode` and if it is
`autoAccept`, then we should eagerly register the protocol handler.
We do this by updating the `ProtocolRegistry` in the resource thread,
where we need to clone it (since there can be ongoing fetches). To
be able to clone handlers, we need a new method on the trait.

Part of #40615

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>