html: Use WeakRef for the `MediaSession` media element (#40875)
The media session is allowed to control the state of the media element
(play, pause) and is holding DOM reference, so the latest registered
media element is not be able to be GCed, so let's use the weak reference
to break this dependency.

Note that `weakReference` property was added to interfaces
- HTMLMediaElement: The `Mutable/WeakRef` requires T to be bound by
`WeakReferenceable`
- HTMLAudio/VideoElement: The `DOM_WEAK_SLOT` must be reserved by
concrete DOM JS class

Testing: No expected testing results changes

The media session UI functionality is implemented for Android and didn't
test due lack of testing sample.

Fixes: https://github.com/servo/servo/issues/40854

Signed-off-by: Andrei Volykhin <andrei.volykhin@gmail.com>