ci: Remove self-hosted runner timeout jobs (#40876)
the runner timeout jobs are a helpful safeguard against bugs (or github
infra failures) that cause jobs to not get picked up by self-hosted
runners, but all they do is wait two minutes and cancel the run,
clogging up our github-hosted runner capacity (limited to 60).

this patch removes the timeout jobs, based on the expectation that our
CI system is reliable enough to not need that automated safeguard.

if you have a build that gets stuck on “Waiting for a runner to pick up
this job...”, it may be a bug or a github infra failure. you can cancel
the build yourself, but please take a screenshot and report it first!

Testing:
- [mach try windows linux
lint](https://github.com/servo/servo/actions/runs/19659755273)

Signed-off-by: Delan Azabani <dazabani@igalia.com>