CI: Add retries for unit-tests with code-coverage
Some unit-tests are flaky, so we rerun them normally.
We didn't do this until now when measuring coverage, which has been
a source of failures for the coverage job.
Additionally we can also remove the install lld step, since we are now
using a newer version of Rust, where we can just use lld out of the box.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>