script: Stop assuming that previous dirty root is still connected (#40922)
We were trying to update the dirty root to be the common ancestor of the
old dirty root and the element noted with dirty descendants. But that
would panic if when the old dirty root had been disconnected.

In that case, just set the dirty root to be the element, as if the old
dirty root was None.

Testing: Adding a crashtest, and one existing test now fails instead of
crashing
Fixes: #40920

Signed-off-by: Oriol Brufau <obrufau@igalia.com>