servoshell: Add support for opening new windows in desktop
This change finishes adding support for opening new windows in the
desktop version of servoshell. A new button is added to the tab bar
which opens a new window (this can be adjusted later, if necessary).
User interface commands now need to be processed in the context of the
`App` as we need access to a reference to the `ActiveEventLoop` to
create a new window.

Fixes #13997.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>