servoshell: Schedule keyboard shortcut exits so that they can be done cleanly (#40933)
This is a speculative fix for a crash mentioned by @mukilan[^1]. It will
also make moving to a more consistent shutdown model (even one driven by
dropping of the `Servo` instance) possible.

[^1]: https://github.com/servo/servo/pull/40883#issuecomment-3584090609

Testing: It's difficult to test this crash because it seems to only
happen on
nixOS and we currently don't have a good way to test crashes on exit.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>