libservo: Clean up Servo exports and export more at the root
This change makes it easier for embedders to embed the types that they
need by exporting almost everything necessary to use Servo at the root
of libservo, apart from a few exceptions. In addition, the `Servo` is moved
to its own file so that public exports can be more easily spotted from
`components/servo/lib.rs`.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>