script: Reverse retargeting algorithm for ShadowRoot.element[s]FromPoint. (#40975)
This implementation is not actually specified (see #40973 and
https://github.com/w3c/csswg-drafts/issues/556), but the implementation
that we have is backwards. Other retargeting uses that describe
"retarget _something_ against _another thing_" map to
`something.retarget(another_thing)`, but what the current code was doing
for element[s]FromPoint was the equivalent of "retarget _this_ against
each element", which is obviously wrong.

Testing: Newly-passing tests.

Signed-off-by: Josh Matthews <josh@joshmatthews.net>