libservo: Finish the `SimpleDialog` API
Finish exposing the `SimpleDialog` API in libservo, abstracting away the
response senders into structures with methods. In addition, add a
default behavior so that embedders do not have to explicitly send a
reponse when dialogs are ignored. This makes the API much harder to
misuse.

Finally, the WebDriver API for setting querying the prompt dialog entry
field is corrected to actually get that value rather than the message.
This makes sense as part of this change as the type safety of the
structures makes setting and getting the message (the old, wrong
behavior) harder to implement.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>