servoshell: Explicitly add the  `Win32_System_Console` feature to `windows-sys` (#40980)
In #40961, we start to use `windows_sys::Win32::System::Console;`. This
is gated behind `Win32_System_Console` feature. We didn't have the
problem merging that one, because it was luckily enabled by some other
dependencies. But @dependabot can act weird: #40979, which would break
the build from time to time.

See
https://github.com/servo/servo/actions/runs/19816434323/job/56768628158.

Fixes: Treacherous bug that is only revealed when other crates's
`windows-sys` is downgraded.

---------

Signed-off-by: Euclid Ye <yezhizhenjiakang@gmail.com>