net: Allow servo to use a http proxy without authentication (#40941)
Servo can now use a http proxy to connect to the internet. This
currently only supports using all connections through the proxy.

This uses the already used hyper_util crate which had the functionality
in it already.
We also switch from tower_service to tower with only util feature flag.
This might increase the binary size slightly.

Testing: Tested on a local machine which is behind a proxy with and
without the flag and it works correctly.

---------

Signed-off-by: Narfinger <Narfinger@users.noreply.github.com>