layout: Handle stacking context tree outlines during display list building (#40984)
Instead of eagerly adding outlines to their own stacking context tree
section, handle them as part of display list building. This means that
the stacking context tree does not need to be rebuilt when outlines on
the page change.

Testing: This is a speculative performance fix, behavior is covered by
existing tests.
Fixes: #40957.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Oriol Brufau <obrufau@igalia.com>