script: Don't dirty with ContentOrHeritage damage when changing state (#40902)
When for example changing the hover state of an element, we were
dirtying it with ContentOrHeritage damage, which forced it to be laid
out again.

This was added in #39102 in order to fix #38989, but for the most part
it's no longer needed.

However, we will now dirty when a `<textarea>` or `<input>` which can be
selected gets or loses focus. This is to ensure that the caret or
selections get correctly updated.

Testing: Manually tested that #38989 is not regressing

Signed-off-by: Oriol Brufau <obrufau@igalia.com>
Co-authored-by: Martin Robinson <mrobinson@igalia.com>